fix: fixed the problem of quiz visibility

+++ hinghwa-dict-backend/quiz/views.py
@@ -10,32 +10,44 @@ from utils.exception.types.bad_request import BadRequestException
 from utils.exception.types.common import CommonException
 from utils.exception.types.not_found import QuizNotFoundException
 from utils.exception.types.unauthorized import UnauthorizedException
+from utils.exception.types.forbidden import ForbiddenException
 from utils.token import token_pass, token_user
 
 
 class SingleQuiz(View):
     # QZ0101 获取单个测试
     def get(self, request, id) -> JsonResponse:
+        quiz = Quiz.objects.filter(id=id)
+        if not quiz.exists():  # 404
+            raise QuizNotFoundException()
+        quiz = quiz[0]
         try:
-            quiz = Quiz.objects.filter(id=id)
-            if not quiz.exists():  # 404
-                raise QuizNotFoundException()
-            quiz = quiz[0]
+            token = token_pass(request.headers)
+            user = token_user(token)
+        except UnauthorizedException:
             if not quiz.visibility:
                 raise QuizNotFoundException()
             return JsonResponse({"quiz": quiz_all(quiz)}, status=200)
-        except CommonException as e:  # 500
-            raise e
+        if (
+            not quiz.visibility
+            and not user.is_superuser
+            and not user.id == quiz.author.id
+        ):
+            raise QuizNotFoundException()
+        return JsonResponse({"quiz": quiz_all(quiz)}, status=200)
 
     # QZ0103 修改单个测试
     def put(self, request, id) -> JsonResponse:
         try:
             body = demjson.decode(request.body)
-            token = token_pass(request.headers, -1)
+            token = token_pass(request.headers)
+            user = token_user(token)
             quiz = Quiz.objects.filter(id=id)
             if not quiz.exists():  # 404
                 raise QuizNotFoundException()
             quiz = quiz[0]
+            if not user.is_superuser and user.id != quiz.author.id:
+                raise ForbiddenException()
             body = body["quiz"]
             for key in body:
                 setattr(quiz, key, body[key])
@@ -49,10 +61,13 @@ class SingleQuiz(View):
     def delete(self, request, id) -> JsonResponse:
         try:
             quiz = Quiz.objects.filter(id=id)
-            token = token_pass(request.headers, -1)
+            token = token_pass(request.headers)
+            user = token_user(token)
             if not quiz.exists():  # 404
                 raise QuizNotFoundException()
             quiz = quiz[0]
+            if not user.is_superuser and user.id != quiz.author.id:
+                raise ForbiddenException()
             quiz.delete()
             return JsonResponse({}, status=200)
         except CommonException as e:  # 500
@@ -62,21 +77,27 @@ class SingleQuiz(View):
 class MultiQuiz(View):
     # QZ0201 搜索测试题
     def get(self, request) -> JsonResponse:
-        result = []
+        quizzes = []
         try:
             quizzes = Quiz.objects.all()
             # keywords此处默认对question进行模糊搜索
             if "keywords" in request.GET:
                 quizzes = quizzes.filter(question__contains=request.GET["keywords"])
             quizzes = list(quizzes)
-            result = [quiz for quiz in quizzes if quiz.visibility]
+
             token = token_pass(request.headers)
             user = token_user(token)
             if not user.is_superuser:
+                result = [
+                    quiz_all(quiz)
+                    for quiz in quizzes
+                    if quiz.visibility or user.id == quiz.author.id
+                ]
                 return JsonResponse({"result": result}, status=200)
             result = [quiz_all(quiz) for quiz in quizzes]
             return JsonResponse({"result": result}, status=200)
         except UnauthorizedException:
+            result = [quiz_all(quiz) for quiz in quizzes if quiz.visibility]
             return JsonResponse({"result": result}, status=200)
         except CommonException as e:  # 500
             raise e
@@ -86,11 +107,13 @@ class MultiQuiz(View):
     # QZ0102 增加单个测试
     def post(self, request) -> JsonResponse:
         body = demjson.decode(request.body)
-        token = token_pass(request.headers, -1)
+        token = token_pass(request.headers)
+        user = token_user(token)
         quiz_form = QuizForm(body)
         if not quiz_form.is_valid():
             raise BadRequestException()
         quiz = quiz_form.save(commit=False)
+        quiz.author = user
         quiz.save()
         return JsonResponse({"quiz": quiz_all(quiz)}, status=200)
